Create markovStateFactory.py
Uses moving averages, relative strength, and volatility metrics to categorize each bar as one of three states: upward trending, downward trending, or flat. 

Upward trending is defined as a the price is generally rising over a period of greater than 15 days. 

Downward trending is defined as a the price is generally decreasing over a period of greater than 15 days. 

Flat is defined as a the price is generally remaining within a stable range over a period of greater than 15 days.

requires changes to fit with the markov state determination modules